Test optimization function

test_optim() function is useful to visualize how optimizers solve the minimization problem by showing the convergence path using a test function. User can choose any test optimization functions provided by torchopt:

"beale", "booth", "bukin_n6", "easom", "goldstein_price", "himmelblau", "levi_n13", "matyas", "rastrigin", "rosenbrock", and "sphere".

Besides these functions, users can pass any function that receives two numerical values and returns a scalar.

Optimization functions are useful to evaluate characteristics of optimization algorithms, such as convergence rate, precision, robustness, and performance. These functions give an idea about the different situations that optimization algorithms can face.

Function test_function() plot the 2D-space of a test optimization function.

test_function(
  test_fn,
  ...,
  bg_x_lim = c(-5, 5),
  bg_y_lim = c(-5, 5),
  bg_xy_breaks = 100,
  bg_z_breaks = 32,
  bg_palette = "viridis",
  ct_levels = 10,
  ct_labels = TRUE,
  ct_color = "#FFFFFF7F"
)

test_optim(
  optim,
  ...,
  opt_hparams = list(lr = 0.01),
  test_fn = "beale",
  steps = 200,
  pt_start_color = "#5050FF7F",
  pt_end_color = "#FF5050FF",
  ln_color = "#FF0000FF",
  ln_weight = 2,
  bg_xy_breaks = 100,
  bg_z_breaks = 32,
  bg_palette = "viridis",
  ct_levels = 10,
  ct_labels = FALSE,
  ct_color = "#FFFFFF7F",
  plot_each_step = FALSE
)

test_optim_valid(
  optim,
  opt_hparams = list(lr = 0.01),
  test_fn = "beale",
  steps = 100
)

Arguments

test_fn

A test function (default "beale").

...

Additional parameters (passed to image function).

bg_x_lim

Background X limits (default NULL)

bg_y_lim

Background Y limits (default NULL)

bg_xy_breaks

Background X and Y resolution (default 100)

bg_z_breaks

Background Z resolution (default 32)

bg_palette

Background palette (default "viridis")

ct_levels

Contour levels (default 10)

ct_labels

Should show contour labels? (default FALSE)

ct_color

Contour color (default "#FFFFFF7F")

optim

Torch optimizer function.

opt_hparams

A list with optimizer initialize parameters (default list(lr = 0.01)).

steps

Number of steps to run (default 100).

pt_start_color

Starting point color (default "#5050FF7F")

pt_end_color

Ending point color (default "#FF5050FF")

ln_color

Line path color (default "#FF0000FF")

ln_weight

Line path weight (default 2)

plot_each_step

Should output each step? (default FALSE)

Author

Rolf Simoes, rolf.simoes@inpe.br